requirements.txt
asgiref==3.2.9
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
django-countries==6.1.2
django-crispy-forms==1.9.1
django-filter==2.3.0
django-widget-tweaks==1.4.8
GDAL==3.1.2
idna==2.9
Pillow==7.1.2
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
stripe==2.48.0
urllib3==1.25.9
